Move diamondsquare to utils
It is used to create a new world, but it's only a tool used by other
actors.
